requirements.txt
colorlog==6.8.2
einops==0.8.0
numpy==1.25.0
opencv_python==4.9.0.80
Pillow==9.4.0
PyYAML==6.0.1
torch==2.0.1
torchmetrics==1.3.1
torchvision==0.15.2
tqdm==4.65.0
tensorboard==2.17.0